refactor(tests): extract `FeedImageDataLoaderSpy` into shared scope to remove duplication
